+++ scripts/Token.js
@@ -23,7 +23,8 @@ function _onDragLeftStart(wrapped, event) {
 
   // If Token Ruler, start a ruler measurement.
   if ( !Settings.get(Settings.KEYS.TOKEN_RULER.ENABLED) ) return;
-  canvas.controls.ruler._onDragStart(event);
+
+  canvas.controls.ruler._onDragStart(event, { isTokenDrag: true });
 }
 
 /**
@@ -109,7 +110,7 @@ function updateToken(document, changes, _options, _userId) {
 
   const ruler = canvas.controls.ruler;
   if ( ruler.active && ruler.token === token ) token._lastMoveDistance = ruler.totalMoveDistance;
-  else token._lastMoveDistance = Ruler.measureMoveDistance(token.position, token.document, { token }).moveDistance;
+  else token._lastMoveDistance = Ruler.measureMoveDistance(token.position, token.document._source, { token }).moveDistance;
   if ( game.combat?.started ) {
     // Store the combat move distance and the last round for which the combat move occurred.
     // Map to each unique combat.